COM-3819: manon's review and resize
